--- x/bep3/types/status.go
@@ -1,54 +1,0 @@
-package types
-
-import (
-	"encoding/json"
-)
-
-type SwapStatus byte
-
-const (
-	NULL      SwapStatus = 0x00
-	Open      SwapStatus = 0x01
-	Completed SwapStatus = 0x02
-	Expired   SwapStatus = 0x03
-)
-
-func NewSwapStatusFromString(str string) SwapStatus {
-	switch str {
-	case "Open", "open":
-		return Open
-	case "Completed", "completed":
-		return Completed
-	case "Expired", "expired":
-		return Expired
-	default:
-		return NULL
-	}
-}
-
-func (status SwapStatus) String() string {
-	switch status {
-	case Open:
-		return "Open"
-	case Completed:
-		return "Completed"
-	case Expired:
-		return "Expired"
-	default:
-		return "NULL"
-	}
-}
-
-func (status SwapStatus) MarshalJSON() ([]byte, error) {
-	return json.Marshal(status.String())
-}
-
-func (status *SwapStatus) UnmarshalJSON(data []byte) error {
-	var s string
-	err := json.Unmarshal(data, &s)
-	if err != nil {
-		return err
-	}
-	*status = NewSwapStatusFromString(s)
-	return nil
-}
